DOMA-2460 Move cache to cache.js && Add .env setting
