+++ apps/condo/index.js
@@ -1,4 +1,4 @@
-const { identity, get } = require('lodash')
+const { identity } = require('lodash')
 const { v4 } = require('uuid')
 const express = require('express')
 const bodyParser = require('body-parser')
@@ -23,12 +23,15 @@ const { formatError } = require('@condo/domains/common/utils/apolloErrorFormatte
 const { hasValidJsonStructure } = require('@condo/domains/common/utils/validation.utils')
 const { SbbolRoutes } = require('@condo/domains/organization/integrations/sbbol/routes')
 const FileAdapter = require('@condo/domains/common/utils/fileAdapter')
+const { KeystoneCacheMiddleware, initCache } = require('@core/keystone/cache')
 const { expressErrorHandler } = require('@condo/domains/common/utils/expressErrorHandler')
 const { OIDCMiddleware } = require('@condo/domains/user/oidc')
 
 
 const IS_ENABLE_DD_TRACE = conf.NODE_ENV === 'production'
 const IS_ENABLE_APOLLO_DEBUG = conf.NODE_ENV === 'development' || conf.NODE_ENV === 'test'
+
+const IS_ENABLE_CACHE = conf.ENABLE_CACHE === '1'
 
 let IS_BUILD_PHASE = conf.PHASE === 'build'
 if (IS_BUILD_PHASE) {
@@ -69,130 +72,9 @@ const keystone = new Keystone({
     },
 })
 
-const originalCreateList = keystone.createList
-
-/**
- * example:
- * {
- *      <requestId> : {
- *          <requestKey>
- *      }
- * }
- * @type {{}}
- */
 const cache = {}
-
-const generateRequestKey = (gqlName, args) => (
-    `${gqlName}-${JSON.stringify(args)}`
-)
-
-const getRequestIdFromContext = (context) => {
-    return get(context, ['req', 'headers', 'x-request-id'], null)
-}
-
-keystone.createList = async (...args) => {
-    const list = originalCreateList.apply(keystone, args)
-
-    // Ensure that if there is a mutation -- we delete cache item!
-    const originalCreateMutation = list.createMutation
-    list.createMutation = async ( data, context, mutationState ) => {
-        const requestId = getRequestIdFromContext(context)
-        if (requestId) { delete cache[requestId] }
-        return await originalCreateMutation.call( list, data, context, mutationState )
-    }
-
-    const originalCreateManyMutation = list.createManyMutation
-    list.createManyMutation = async ( data, context, mutationState ) => {
-        const requestId = getRequestIdFromContext(context)
-        if (requestId) { delete cache[requestId] }
-        return await originalCreateManyMutation.call( list, data, context, mutationState )
-    }
-
-    // For some reason mutationState == context in case of update. Ask Keystone!
-    const originalUpdateMutation = list.updateMutation
-    list.updateMutation = async ( data, context, mutationState ) => {
-        const requestId = getRequestIdFromContext(mutationState)
-        if (requestId) { delete cache[requestId] }
-        return await originalUpdateMutation.call( list, data, context, mutationState )
-    }
-
-    const originalUpdateManyMutation = list.updateManyMutation
-    list.updateManyMutation = async ( data, context, mutationState ) => {
-        const requestId = getRequestIdFromContext(mutationState)
-        if (requestId) { delete cache[requestId] }
-        return await originalUpdateManyMutation.call( list, data, context, mutationState )
-    }
-
-    const originalDeleteMutation = list.deleteMutation
-    list.deleteMutation = async ( data, context, mutationState ) => {
-        const requestId = getRequestIdFromContext(mutationState)
-        if (requestId) { delete cache[requestId] }
-        return await originalDeleteMutation.call( list, data, context, mutationState )
-    }
-
-    const originalListQuery = list.listQuery
-    list.listQuery = async function (args, context, gqlName, info, from) {
-
-        let key = null
-        const requestId = getRequestIdFromContext(context)
-
-        if (requestId) {
-            key = generateRequestKey(gqlName, args)
-
-            if (!(requestId in cache)) {
-                cache[requestId] = {}
-            }
-
-            // Drop the key, if the operation type is mutation
-            const operationType = get(info, ['operation', 'operation'])
-            if (operationType !== 'query') {
-                delete cache[requestId][key]
-            }
-
-            if (key in cache[requestId]) {
-                return cache[requestId][key]
-            }
-        }
-
-        const listResult = await originalListQuery.call(list, args, context, gqlName, info, from)
-
-        if (requestId && key) { cache[requestId][key] = listResult }
-
-        return listResult
-    }
-
-    const originalItemQuery = list.itemQuery
-    list.itemQuery = async (args, context, gqlName, info, from) => {
-
-        let key = null
-        const requestId = getRequestIdFromContext(context)
-
-        if (requestId) {
-            key = generateRequestKey(gqlName, args)
-
-            if (!(requestId in cache)) {
-                cache[requestId] = {}
-            }
-
-            // Drop the key, if the operation type is mutation
-            const operationType = get(info, ['operation', 'operation'])
-            if (operationType !== 'query') {
-                delete cache[requestId][key]
-            }
-
-            if (key in cache[requestId]) {
-                return cache[requestId][key]
-            }
-        }
-
-        const itemQuery = await originalItemQuery.call(list, args, context, gqlName, info, from)
-
-        if (requestId && key) { cache[requestId][key] = itemQuery }
-
-        return itemQuery
-    }
-
-    return list
+if (IS_ENABLE_CACHE) {
+    initCache(keystone, cache)
 }
 
 if (!IS_BUILD_PHASE) {
@@ -260,33 +142,12 @@ class CustomBodyParserMiddleware {
     }
 }
 
-class KeystoneCacheMiddleware {
-    prepareMiddleware ({ keystone, dev, distDir }) {
-        const app = express()
-        app.use((req, res, next) => {
-            console.log(`${req.method} ${req.originalUrl} [STARTED]`)
-
-            const requestId = get()
-
-            // res.on('finish', () => {
-            //     console.log(`${req.method} ${req.originalUrl} [FINISHED]`)
-            // })
-
-            res.on('close', () => {
-                console.log('lol')
-            })
-
-            next()
-        })
-    }
-}
-
 module.exports = {
     keystone,
     apps: [
         new OIDCMiddleware(),
         new CustomBodyParserMiddleware(),
-        new KeystoneCacheMiddleware(),
+        new KeystoneCacheMiddleware(cache),
         new GraphQLApp({
             apollo: {
                 formatError,
